debuged for python 3 and added jupyter notebook

+++ rn/model/checkerboard.py
@@ -7,21 +7,21 @@ from rn.rsf.model import Model
 def checkerboard( nx, nz, nx_checker, nz_checker ) :
 
   # first create slightly larger checke3rboard patterns
-  nchecker_x = nx / nx_checker / 2 
-  nchecker_z = nz / nz_checker / 2
+  nchecker_x = int( nx / nx_checker / 2  )
+  nchecker_z = int( nz / nz_checker / 2 )
 
   print( nchecker_x, nchecker_z )
 
   if nchecker_x * 2 <  nx :
     nchecker_x += 1 
     print(  nchecker_x )
-    ix0 = ( nchecker_x *nx_checker * 2- nx ) / 2 
+    ix0 = int( ( nchecker_x *nx_checker * 2- nx ) / 2  )
   else :
     ix0 = 0 
  
   if nchecker_z *2 <  nz :
     nchecker_z += 1
-    iz0 = ( nchecker_z * nz_checker * 2- nz ) / 2
+    iz0 = int( ( nchecker_z * nz_checker * 2- nz ) / 2 )
   else :
     iz0 = 0
